+++ app/models/asyncapi/client/job.rb
@@ -71,7 +71,6 @@ module Asyncapi::Client
       args[:time_out_at] = time_out.from_now if time_out
       job = create(args)
       JobPostWorker.perform_async(job.id, url)
-      CleanerWorker.perform_async
     end
 
     def url