+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- Cleaner worker is scheduled to run every day instead of every time a job is created
+
 ## [0.5.1] - 2016-05-18
 ### Fixed
 - [Do not include redis immediately](https://github.com/G5/asyncapi-client/pull/20)